handled errer cases with try catch Relates #56
